+++ apps/test/src/domain/iam/services/authorization/authorization.service.ts
@@ -5,16 +5,23 @@ import { AuthorizationData, IAMAuthorizationService as BaseAuthorizationService 
 
 import { AuthorizationPoint, CacheAuthorizationPointsEntityService } from '../../../../data/cache';
 import { AuthorizationPointsService as DBUAuthorizationPointsEntityService } from '../../../../data/db';
+import { IAMTokenManagerService } from '../tokenManager';
 
 @Injectable()
 export class IAMAuthorizationService extends BaseAuthorizationService<AuthorizationPoint> {
   constructor(
     protected dataAuthorizationPointsService: CacheAuthorizationPointsEntityService,
-    protected dataDBAuthorizationPointsService: DBUAuthorizationPointsEntityService
+    protected dataDBAuthorizationPointsService: DBUAuthorizationPointsEntityService,
+    protected tokenManager: IAMTokenManagerService
   ) {
-    super(dataAuthorizationPointsService, [DomainMethod.Find], {
-      db: dataDBAuthorizationPointsService as DataEntityService<Partial<AuthorizationPoint>>
-    });
+    super(
+      dataAuthorizationPointsService,
+      [DomainMethod.Find],
+      {
+        db: dataDBAuthorizationPointsService as DataEntityService<Partial<AuthorizationPoint>>
+      },
+      tokenManager
+    );
   }
 
   async mapAuthorizationPoints(moduleName: string): Promise<AuthorizationData<unknown>> {