+++ packages/api-http/src/middlewares/http.middlewares.authorization.ts
@@ -28,7 +28,7 @@ export class HTTPAuthorizationMiddleware<User extends object> implements NestMid
     @Inject(Constants.API_MODULE_AUTHORIZATION_SERVICE)
     // eslint-disable-next-line no-unused-vars
     protected authorizationService: IAMAuthorizationService<AuthorizationPoint<unknown>>,
-    @Inject(Constants.AUTORIZATION_MIDDLEWARE_TOKEN_MANAGER_SERVICE)
+    @Inject(Constants.AUTHORIZATION_MIDDLEWARE_TOKEN_MANAGER_SERVICE)
     // eslint-disable-next-line no-unused-vars
     protected tokenManager?: IAMTokenManagerService<UserTokenEnityFields>,
     @Inject(Constants.AUTHENTICATION_MIDDLEWARE_USERS_SERVICE)