+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.0-alpha63
+- Common
+- `packages/domain-iam`
+  - Typing and typo fixes for the authorization service.
+
 # 1.0.0-alpha62
 - Common
   - BREAKING: Renamed all "persistance" packages whose name is confusing, misleading, factually inaccurate and - worst of all - contains a typo. They're now "data" packages.